refactor: move FAQ and About pages to /public, add MainLayout
