--- learnhub-ui/src/pages/About.tsx
@@ -1,304 +1,0 @@
-export default function About() {
-    return (
-        <div className="page-content bg-white">
-            <div
-                className="page-banner ovbl-dark"
-                style={{ backgroundImage: "url(/assets/images/banner/banner3.jpg)" }}>
-                <div className="container">
-                    <div className="page-banner-entry">
-                        <h1 className="text-white">About Us 2</h1>
-                    </div>
-                </div>
-            </div>
-            <div className="breadcrumb-row">
-                <div className="container">
-                    <ul className="list-inline">
-                        <li>
-                            <a href="#">Home</a>
-                        </li>
-                        <li>About Us 2</li>
-                    </ul>
-                </div>
-            </div>
-            <div className="content-block">
-                <div className="section-area section-sp1">
-                    <div className="container">
-                        <div className="row">
-                            <div className="col-lg-6 m-b30">
-                                <h2 className="title-head ">
-                                    Learn a new skill online
-                                    <br /> <span className="text-primary"> on your time</span>
-                                </h2>
-                                <h4>
-                                    <span className="counter">57,000 </span> Online Courses
-                                </h4>
-                                <p>
-                                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem
-                                    Ipsum has been the industry's standard dummy text ever since the 1500s, when an
-                                    unknown printer took a galley of type.
-                                </p>
-                                <a href="#" className="btn button-md">
-                                    Join Now
-                                </a>
-                            </div>
-                            <div className="col-lg-6">
-                                <div className="row">
-                                    <div className="col-lg-6 col-md-6 col-sm-6 m-b30">
-                                        <div className="feature-container">
-                                            <div className="feature-md text-white m-b20">
-                                                <a href="#" className="icon-cell">
-                                                    <img src="/assets/images/icon/icon1.png" alt="" />
-                                                </a>
-                                            </div>
-                                            <div className="icon-content">
-                                                <h5 className="ttr-tilte">Our Philosophy</h5>
-                                                <p>Lorem ipsum dolor sit amet, consectetuer adipiscing.</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                    <div className="col-lg-6 col-md-6 col-sm-6 m-b30">
-                                        <div className="feature-container">
-                                            <div className="feature-md text-white m-b20">
-                                                <a href="#" className="icon-cell">
-                                                    <img src="/assets/images/icon/icon2.png" alt="" />
-                                                </a>
-                                            </div>
-                                            <div className="icon-content">
-                                                <h5 className="ttr-tilte">Kingster's Principle</h5>
-                                                <p>Lorem ipsum dolor sit amet, consectetuer adipiscing.</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                    <div className="col-lg-6 col-md-6 col-sm-6 m-b30">
-                                        <div className="feature-container">
-                                            <div className="feature-md text-white m-b20">
-                                                <a href="#" className="icon-cell">
-                                                    <img src="/assets/images/icon/icon3.png" alt="" />
-                                                </a>
-                                            </div>
-                                            <div className="icon-content">
-                                                <h5 className="ttr-tilte">Key Of Success</h5>
-                                                <p>Lorem ipsum dolor sit amet, consectetuer adipiscing.</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                    <div className="col-lg-6 col-md-6 col-sm-6 m-b30">
-                                        <div className="feature-container">
-                                            <div className="feature-md text-white m-b20">
-                                                <a href="#" className="icon-cell">
-                                                    <img src="/assets/images/icon/icon4.png" alt="" />
-                                                </a>
-                                            </div>
-                                            <div className="icon-content">
-                                                <h5 className="ttr-tilte">Our Philosophy</h5>
-                                                <p>Lorem ipsum dolor sit amet, consectetuer adipiscing.</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div className="section-area bg-gray section-sp2 choose-bx">
-                    <div className="container">
-                        <div className="row">
-                            <div className="col-md-12 heading-bx text-center">
-                                <h2 className="title-head text-uppercase m-b0">
-                                    Why Choose <span> Our Institution</span>
-                                </h2>
-                                <p>
-                                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem
-                                    Ipsum has been the industry's.
-                                </p>
-                            </div>
-                        </div>
-                        <div className="row choose-bx-in">
-                            <div className="col-lg-4 col-md-4 col-sm-6">
-                                <div className="service-bx">
-                                    <div className="action-box">
-                                        <img src="/assets/images/our-services/pic1.jpg" alt="" />
-                                    </div>
-                                    <div className="info-bx text-center">
-                                        <div className="feature-box-sm radius bg-white">
-                                            <i className="fa fa-bank text-primary"></i>
-                                        </div>
-                                        <h4>
-                                            <a href="#">Best Industry Leaders</a>
-                                        </h4>
-                                        <a href="#" className="btn radius-xl">
-                                            View More
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="col-lg-4 col-md-4 col-sm-6">
-                                <div className="service-bx">
-                                    <div className="action-box">
-                                        <img src="/assets/images/our-services/pic2.jpg" alt="" />
-                                    </div>
-                                    <div className="info-bx text-center">
-                                        <div className="feature-box-sm radius bg-white">
-                                            <i className="fa fa-book text-primary"></i>
-                                        </div>
-                                        <h4>
-                                            <a href="#">Learn Courses Online</a>
-                                        </h4>
-                                        <a href="#" className="btn radius-xl">
-                                            View More
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="col-lg-4 col-md-4 col-sm-12">
-                                <div className="service-bx m-b0">
-                                    <div className="action-box">
-                                        <img src="/assets/images/our-services/pic3.jpg" alt="" />
-                                    </div>
-                                    <div className="info-bx text-center">
-                                        <div className="feature-box-sm radius bg-white">
-                                            <i className="fa fa-file-text-o text-primary"></i>
-                                        </div>
-                                        <h4>
-                                            <a href="#">Book Library & Store</a>
-                                        </h4>
-                                        <a href="#" className="btn radius-xl">
-                                            View More
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div className="section-area content-inner section-sp1">
-                    <div className="container">
-                        <div className="section-content">
-                            <div className="row">
-                                <div className="col-lg-3 col-md-6 col-sm-6 col-6 m-b30">
-                                    <div className="counter-style-1">
-                                        <div className="text-primary">
-                                            <span className="counter">3000</span>
-                                            <span>+</span>
-                                        </div>
-                                        <span className="counter-text">Completed Projects</span>
-                                    </div>
-                                </div>
-                                <div className="col-lg-3 col-md-6 col-sm-6 col-6 m-b30">
-                                    <div className="counter-style-1">
-                                        <div className="text-black">
-                                            <span className="counter">2500</span>
-                                            <span>+</span>
-                                        </div>
-                                        <span className="counter-text">Happy Clients</span>
-                                    </div>
-                                </div>
-                                <div className="col-lg-3 col-md-6 col-sm-6 col-6 m-b30">
-                                    <div className="counter-style-1">
-                                        <div className="text-primary">
-                                            <span className="counter">1500</span>
-                                            <span>+</span>
-                                        </div>
-                                        <span className="counter-text">Questions Answered</span>
-                                    </div>
-                                </div>
-                                <div className="col-lg-3 col-md-6 col-sm-6 col-6 m-b30">
-                                    <div className="counter-style-1">
-                                        <div className="text-black">
-                                            <span className="counter">1000</span>
-                                            <span>+</span>
-                                        </div>
-                                        <span className="counter-text">Ordered Coffee's</span>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div className="section-area bg-gray section-sp1 our-story">
-                    <div className="container">
-                        <div className="row align-items-center d-flex">
-                            <div className="col-lg-5 col-md-12 heading-bx">
-                                <h2 className="m-b10">Our Story</h2>
-                                <h5 className="fw4">It is a long established fact that a reade.</h5>
-                                <p>
-                                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem
-                                    Ipsum has been the industry's standard dummy text ever since the 1500s, when an
-                                    unknown printer took a galley of type and scrambled it to make a type specimen book.
-                                    It has survived not only five centuries, but also the leap into electronic
-                                    typesetting, remaining essentially unchanged.
-                                </p>
-                                <a href="#" className="btn">
-                                    Read More
-                                </a>
-                            </div>
-                            <div className="col-lg-7 col-md-12 heading-bx p-lr">
-                                <div className="video-bx">
-                                    <img src="/assets/images/about/pic1.jpg" alt="" />
-                                    <a
-                                        href="https://www.youtube.com/watch?v=x_sJzVe9P_8"
-                                        className="popup-youtube video">
-                                        <i className="fa fa-play"></i>
-                                    </a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div className="section-area section-sp2">
-                    <div className="container">
-                        <div className="row">
-                            <div className="col-md-12 heading-bx left">
-                                <h2 className="title-head text-uppercase">
-                                    what people <span>say</span>
-                                </h2>
-                                <p>
-                                    It is a long established fact that a reader will be distracted by the readable
-                                    content of a page
-                                </p>
-                            </div>
-                        </div>
-                        <div className="testimonial-carousel owl-carousel owl-btn-1 col-12 p-lr0">
-                            <div className="item">
-                                <div className="testimonial-bx">
-                                    <div className="testimonial-thumb">
-                                        <img src="/assets/images/testimonials/pic1.jpg" alt="" />
-                                    </div>
-                                    <div className="testimonial-info">
-                                        <h5 className="name">Peter Packer</h5>
-                                        <p>-Art Director</p>
-                                    </div>
-                                    <div className="testimonial-content">
-                                        <p>
-                                            Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                                            Lorem Ipsum has been the industry's standard dummy text ever since the
-                                            1500s, when an unknown printer took a galley of type...
-                                        </p>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="item">
-                                <div className="testimonial-bx">
-                                    <div className="testimonial-thumb">
-                                        <img src="/assets/images/testimonials/pic2.jpg" alt="" />
-                                    </div>
-                                    <div className="testimonial-info">
-                                        <h5 className="name">Peter Packer</h5>
-                                        <p>-Art Director</p>
-                                    </div>
-                                    <div className="testimonial-content">
-                                        <p>
-                                            Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-                                            Lorem Ipsum has been the industry's standard dummy text ever since the
-                                            1500s, when an unknown printer took a galley of type...
-                                        </p>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-}
