+++ learnhub-ui/src/pages/index.ts
@@ -1,14 +1,23 @@
+// Public pages
 export { default as Landing } from "./public/Landing";
-export { default as Login } from "./auth/Login";
-export { default as ManagerLogin } from "./auth/ManagerLogin";
-export { default as Register } from "./auth/Register";
-export { default as NotFound } from "./error/NotFound";
-export { default as Unauthorized } from "./error/Unauthorized";
-export { default as ForgotPassword } from "./auth/ForgotPassword";
-export { default as ResetPassword } from "./auth/ResetPassword";
-export { default as UserProfile } from "./home/UserProfile";
 export { default as TeacherDetails } from "./public/TeacherDetails";
 export { default as CourseList } from "./public/CourseList";
 export { default as ContactUs } from "./public/ContactUs";
+export { default as FAQ } from "./public/FAQ";
+export { default as About } from "./public/About";
+
+// Auth pages
+export { default as Register } from "./auth/Register";
+export { default as Login } from "./auth/Login";
+export { default as ManagerLogin } from "./auth/ManagerLogin";
+export { default as ForgotPassword } from "./auth/ForgotPassword";
+export { default as ResetPassword } from "./auth/ResetPassword";
+
+// User home
 export { default as Home } from "./home/Home";
+export { default as UserProfile } from "./home/UserProfile";
 export { default as Mailbox } from "./home/teachermanager/Mailbox";
+
+// Error boundary
+export { default as NotFound } from "./error/NotFound";
+export { default as Unauthorized } from "./error/Unauthorized";
